frontend/requirements.txt
streamlit
requests
plotly
pandas
python-jose[cryptography]
